Add files via upload
/* project-Part1.cpp :
This file contains the 'main' function. Program execution begins and ends there.
The Main Menu: get the user's menu choice. Make sure to validate the choice.
Write the loop so that the program goes back to the go back to main menu after the requested task is completed.
If you have not written the code to complete the requested task in your program,
"This task is not complete" message to the user  for now and go back to displaying the main menu. Later,
you will complete the missing parts of the program and submit with your final project submission.
*/